added an alert

+++ src/components/ShoppingItem/ShoppingItem.jsx
@@ -1,17 +1,34 @@
 import axios from "axios";
 import '../../../src/index.css';
+import Swal from 'sweetalert2';
 
 function ShoppingItem ({item, getSupplies}) {
 
     const deleteItem = () => {
-        axios({
-            method: 'DELETE',
-            url: `/api/shopping-list/delete/${item.id}`
-        })
-        .then(response => {
-            getSupplies();
-        })
-        .catch(err => {
+        Swal.fire({
+            title: "You're about to obliterate this item from your list!!",
+            text: "You won't be able to undo this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#04bb99",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, obliterate it from my shopping list!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                axios({
+                    method: 'DELETE',
+                    url: `/api/shopping-list/${item.id}`
+                })
+                .then(response => {
+                    getSupplies();
+                })
+              Swal.fire({
+                title: "Obliterated!!",
+                text: "Your item has been obliterated from your shopping list!",
+                icon: "success"
+              });
+            }
+          }).catch(err => {
             alert('Error Deleting Item')
             console.log(err);
         })
